Improved logging and partial rework of UI
