+++ code/datums/ai/basic_mobs/basic_ai_behaviors/targeting.dm
@@ -74,7 +74,7 @@ GLOBAL_LIST_INIT(target_interested_atoms, typecacheof(list(/mob, /obj/machinery/
 
 	var/atom/target = pick_final_target(controller, filtered_targets)
 
-	EVLOG_MAPTEXT(controller, "AI Targetting", "[controller.pawn] has selected [target] as a target for blackboard key [target_key]!", get_turf(target), "Target: [target]")
+	EVLOG_MAPTEXT(controller, "AI Targetting", "[controller.pawn] has selected [target] as a target for blackboard key [target_key]! Behavior: [src]", get_turf(target), "Target: [target]")
 	EVLOG_LINES(controller, "AI Targetting", "Line to target", get_turf(controller.pawn), get_turf(target))
 
 	controller.set_blackboard_key(target_key, target)
