+++ code/datums/ai/generic/find_and_set.dm
@@ -15,7 +15,7 @@
 	if(isnull(find_this_thing))
 		return AI_BEHAVIOR_DELAY | AI_BEHAVIOR_FAILED
 
-	EVLOG_MAPTEXT(controller, "AI Targetting", "[controller.pawn] has selected [find_this_thing] as a target for blackboard key [set_key]!", get_turf(find_this_thing), "Target: [find_this_thing]")
+	EVLOG_MAPTEXT(controller, "AI Targetting", "[controller.pawn] has selected [find_this_thing] as a target for blackboard key [set_key]! Behavior: [src]", get_turf(find_this_thing), "Target: [find_this_thing]")
 	EVLOG_LINES(controller, "AI Targetting", "Line to target", get_turf(controller.pawn), get_turf(find_this_thing))
 	controller.set_blackboard_key(set_key, find_this_thing)
 	return AI_BEHAVIOR_DELAY | AI_BEHAVIOR_SUCCEEDED
